#68/check authorization on every endpoint

+++ src/controllers/pdsHandbook/pdsHandbook.delete.ts
@@ -1,11 +1,15 @@
 import express from 'express';
 import db from '../../db';
 import { PdsHandbookModel } from '../../models';
+import { checkAuthorization } from '../../utils/authorize';
 
 const del = (req: express.Request, res: express.Response): void => {
   db.connect();
 
   const { chapterId } = req.params;
+
+  const isAuthorized = checkAuthorization(req, res, { checkAdmin: true });
+  if (!isAuthorized) return;
 
   PdsHandbookModel.findOneAndDelete({ chapterId: chapterId })
     .then((chapter) => {
